Add DataType enum and ChunkedIter
Implement methods for DataType and ChunkedIter and add test for iterating through bytes.
ChunkedIter is an iterator that returns n bytes in a vector as a chunk upon next().
Consumes the underlying DataType.

Test for iterating through file not yet written. ChunkedIter.next() should propagate IO errors
in the future.