Implement MD2 algorithm
MD2 algorithm implemented based on the RFC, though the RFC referred was outdated and
was missing an XOR operation during checksum calculation. This implementation only makes
a single pass through the whole data, which means that only a file would be read only
once to calculate the hash, as opposed to the original implementation.